Use C99 designated initializer
Instead of using a non standard feature and since gcc 2.5 depreciated
syntax to specify designated initializer it is better to use the C99
syntax.
Obsoleted and depreciated features are known to get removed without any
further informations in GCC 4.x releases. So the build could break in
one of the next minor or major releases of the compiler suite.
C89 (aka ANSI C) had no equivalent feature so we must push gcc std
version to gnu99 (gnu dialect of c99) instead of gnu89 to get support
for it.

More information about that problem can be found in GCC's manual
"5.23 Designated Initializers".

Signed-off-by: Sven Eckelmann <sven.eckelmann@gmx.de>