batmand: Fix FTBFS on second build attempt
make clean will try to remove all files and folders which end with *.d.
This fails when using the Debian package script on the leftover package
folder with etc/init.d included because the rm is not started in
recursive deletion mode. This find can be replaced with a simpler
statement without using its functionality.

Signed-off-by: Sven Eckelmann <sven@narfation.org>