batmand: Remove subversion/svk revision information
The current version of batmand cannot be obtained from subversion or
svk anymore. It is unnecessary to provide the functionality to retrieve
the current revision using these tools.

We only have to differenciate between git checkout and official release.
We cannot detect a git snapshot and thus the distribution has to provide
the REVISION variable for snapshots.

Signed-off-by: Sven Eckelmann <sven@narfation.org>