batmand: Remove unmaintained batgat kernel module
The batgat kernel module has to be ported to every new kernel version to
be compatible with always moving kernel API. It currently only crashes
when be used with modern kernels.

It doesn't fulfill the requirements to be merged into the upstream
kernel and therefore needs an updated compat layer before it can be
readded to batmand.

Signed-off-by: Sven Eckelmann <sven@narfation.org>
Signed-off-by: Elektra Wagenrad <onelektra@gmx.net>