batmand: Don't cause NULL ptr dereference when searching through HNAs
The _hna_global_add function searches through the list of HNAs to check if
there is already a hna for this orig pointer. Otherwise it is automatically
allocated.

This was implemented by setting hna_orig_ptr to NULL after the comparison. This
will most likely crash the program because the list_for_each_entry
implementation uses hna_orig_ptr to find the next entry before doing the loop
condition check. The next entry cannot be found by dereferencing NULL+epsilon.

Signed-off-by: Sven Eckelmann <sven@narfation.org>
Signed-off-by: Elektra Wagenrad <onelektra@gmx.net>