batmand: Don't try to schedule_own_packet with no if_incoming
send_outstanding_packets checks if a forw_node has a correct if_incoming.
Otherwise it jumps to packet_free to deallocate the packet infrastructure. But
this also schedules packets with the batman interfaces as target incoming_if.
This is known to be NULL but is dereferenced in schedule_own_packet.

This NULL dereference should be avoided.

Signed-off-by: Sven Eckelmann <sven@narfation.org>
Signed-off-by: Elektra Wagenrad <onelektra@gmx.net>